+++ spec/integration/add_id_repository_spec.rb
@@ -1,22 +1,19 @@
 # frozen_string_literal: true
 
 require 'elasticsearch'
-require 'register_sources_oc/repositories/add_id_repository'
+require 'register_sources_oc/repository'
 require 'register_sources_oc/services/es_index_creator'
 require 'register_sources_oc/structs/add_id'
 
-RSpec.describe RegisterSourcesOc::Repositories::AddIdRepository do
-  subject { described_class.new(client: es_client, index:) }
+RSpec.describe RegisterSourcesOc::Repository do
+  subject { described_class.new(RegisterSourcesOc::AddId, client: es_client, index:) }
 
-  let(:index) { SecureRandom.uuid }
+  let(:index) { "tmp-#{SecureRandom.uuid}" }
   let(:es_client) { Elasticsearch::Client.new }
 
   before do
-    index_creator = RegisterSourcesOc::Services::EsIndexCreator.new(
-      add_ids_index: index,
-      client: es_client
-    )
-    index_creator.create_add_ids_index
+    index_creator = RegisterSourcesOc::Services::EsIndexCreator.new(client: es_client)
+    index_creator.create_add_ids_index(index)
   end
 
   after do
