+++ spec/unit/repositories/add_id_repository_spec.rb
@@ -1,9 +1,10 @@
 # frozen_string_literal: true
 
-require 'register_sources_oc/repositories/add_id_repository'
+require 'register_sources_oc/repository'
+require 'register_sources_oc/structs/add_id'
 
-RSpec.describe RegisterSourcesOc::Repositories::AddIdRepository do
-  subject { described_class.new(client: es_client, index:) }
+RSpec.describe RegisterSourcesOc::Repository do
+  subject { described_class.new(RegisterSourcesOc::AddId, client: es_client, index:) }
 
   let(:es_client) { double 'es_client' }
   let(:index) { double 'index' }
